Extract `//cuttlefish/host/commands/assemble_cvd/disk:initialize_instance_composite_disk`
Bug: 420961790
