+++ base/cvd/cuttlefish/host/commands/assemble_cvd/BUILD.bazel
@@ -188,11 +188,11 @@ cc_library(
     deps = [
         "//cuttlefish/common/libs/utils:files",
         "//cuttlefish/common/libs/utils:result",
-        "//cuttlefish/host/commands/assemble_cvd/disk",
         "//cuttlefish/host/commands/assemble_cvd/disk:factory_reset_protected",
         "//cuttlefish/host/commands/assemble_cvd/disk:gem5_image_unpacker",
         "//cuttlefish/host/commands/assemble_cvd/disk:generate_persistent_bootconfig",
         "//cuttlefish/host/commands/assemble_cvd/disk:generate_persistent_vbmeta",
+        "//cuttlefish/host/commands/assemble_cvd/disk:initialize_instance_composite_disk",
         "//cuttlefish/host/commands/assemble_cvd/disk:kernel_ramdisk_repacker",
         "//cuttlefish/host/commands/assemble_cvd:boot_config",
         "//cuttlefish/host/commands/assemble_cvd:boot_image_utils",
