+++ base/cvd/cuttlefish/host/commands/assemble_cvd/disk/BUILD.bazel
@@ -6,48 +6,6 @@ package(
 )
 
 cc_library(
-    name = "disk",
-    srcs = [
-        "initialize_instance_composite_disk.cc",
-    ],
-    hdrs = ["disk.h"],
-    copts = COPTS,
-    strip_include_prefix = "//cuttlefish",
-    deps = [
-        "//cuttlefish/common/libs/fs",
-        "//cuttlefish/common/libs/utils:files",
-        "//cuttlefish/common/libs/utils:result",
-        "//cuttlefish/common/libs/utils:size_utils",
-        "//cuttlefish/common/libs/utils:subprocess",
-        "//cuttlefish/host/commands/assemble_cvd:boot_config",
-        "//cuttlefish/host/commands/assemble_cvd:boot_image_utils",
-        "//cuttlefish/host/commands/assemble_cvd:bootconfig_args",
-        "//cuttlefish/host/commands/assemble_cvd:disk_builder",
-        "//cuttlefish/host/commands/assemble_cvd:vendor_dlkm_utils",
-        "//cuttlefish/host/commands/assemble_cvd/disk:factory_reset_protected",
-        "//cuttlefish/host/commands/assemble_cvd/disk:gem5_image_unpacker",
-        "//cuttlefish/host/commands/assemble_cvd/disk:generate_persistent_bootconfig",
-        "//cuttlefish/host/commands/assemble_cvd/disk:generate_persistent_vbmeta",
-        "//cuttlefish/host/commands/assemble_cvd/disk:kernel_ramdisk_repacker",
-        "//cuttlefish/host/libs/avb",
-        "//cuttlefish/host/libs/config:cuttlefish_config",
-        "//cuttlefish/host/libs/config:data_image",
-        "//cuttlefish/host/libs/config:known_paths",
-        "//cuttlefish/host/libs/feature",
-        "//cuttlefish/host/libs/image_aggregator",
-        "//cuttlefish/host/libs/vm_manager",
-        "//libbase",
-        "@gflags",
-    ],
-)
-
-clang_tidy_test(
-    name = "disk_clang_tidy",
-    srcs = [":disk"],
-    tags = ["clang_tidy", "clang-tidy"],
-)
-
-cc_library(
     name = "factory_reset_protected",
     srcs = ["factory_reset_protected.cc"],
     hdrs = ["factory_reset_protected.h"],
@@ -139,6 +97,31 @@ cc_library(
 clang_tidy_test(
     name = "generate_persistent_vbmeta_clang_tidy",
     srcs = [":generate_persistent_vbmeta"],
+    tags = ["clang_tidy", "clang-tidy"],
+)
+
+cc_library(
+    name = "initialize_instance_composite_disk",
+    srcs = ["initialize_instance_composite_disk.cc"],
+    hdrs = ["initialize_instance_composite_disk.h"],
+    copts = COPTS,
+    strip_include_prefix = "//cuttlefish",
+    deps = [
+        "//cuttlefish/common/libs/utils:files",
+        "//cuttlefish/common/libs/utils:result",
+        "//cuttlefish/host/commands/assemble_cvd/disk:factory_reset_protected",
+        "//cuttlefish/host/commands/assemble_cvd/disk:generate_persistent_vbmeta",
+        "//cuttlefish/host/commands/assemble_cvd:disk_builder",
+        "//cuttlefish/host/libs/config:cuttlefish_config",
+        "//cuttlefish/host/libs/image_aggregator",
+        "//cuttlefish/host/libs/vm_manager",
+        "@gflags",
+    ],
+)
+
+clang_tidy_test(
+    name = "initialize_instance_composite_disk_clang_tidy",
+    srcs = [":initialize_instance_composite_disk"],
     tags = ["clang_tidy", "clang-tidy"],
 )
