+++ base/cvd/cuttlefish/host/commands/assemble_cvd/disk/initialize_instance_composite_disk.h
